Shuffled the instrs
